requirements.dev.txt
# Debugging
django-debug-toolbar==4.2.0
ipdb==0.13.13

# Code quality and formatting
black==23.10.1
flake8==6.1.0
isort==5.12.0
mypy==1.6.1

# Testing
pytest==7.4.3
pytest-django==4.7.0
pytest-cov==4.1.0
factory-boy==3.3.0
faker==19.12.0

# Documentation
Sphinx==7.2.6
sphinx-rtd-theme==1.3.0

# Development server
django-extensions==3.2.3
Werkzeug==3.0.1

# API testing
httpie==3.2.2

# Database visualization
django-silk==5.0.4

# Profiling
pyinstrument==4.6.1

# Mocking
responses==0.23.3

# Improved REPL
ipython==8.16.1

# Environment management
python-dotenv==1.0.0

# Task monitoring (for Celery)
flower==2.0.1